+++ projects/ngrx-rtk-query/src/lib/store-rtk-query.module.ts
@@ -1,12 +1,14 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { setupListeners } from '@reduxjs/toolkit/query';
-import { defaultConfig, StoreQueryConfig, STORE_RTK_QUERY_CONFIG } from './store-rtk-query.config';
+import { defaultConfig, setupConfig, StoreQueryConfig, STORE_RTK_QUERY_CONFIG } from './store-rtk-query.config';
 import { dispatch, ThunkService } from './thunk.service';
 
 @NgModule({})
 export class StoreRtkQueryModule {
   static forRoot(config?: Partial<StoreQueryConfig>): ModuleWithProviders<StoreRtkQueryModule> {
     const moduleConfig = { ...defaultConfig, ...config };
+
+    setupConfig(moduleConfig);
 
     if (moduleConfig.setupListeners) {
       setupListeners(dispatch, moduleConfig.customHandler);